Refactor time parsing functions to a common shared function
Previously there was some code that reordered the list of date
formatters before exhaustively trying each one (stopping on a
successful parse). Now there's code that is similar but for time
parsing. This factors out that common code and parameterizes the
function with a list of date formatters that both can use.